agente-ia-odonto/services/core/requirements.txt
# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.25.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP Client
httpx==0.25.2

# Configuração e ambiente
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilidades
python-dateutil==2.8.2

# Providers de IA
anthropic==0.39.0
openai==1.30.1
google-generativeai==0.7.2

# Logging
python-json-logger==2.0.7